\name{Experiment}
\alias{Experiment}
\title{Experiment constructor}
\usage{
Experiment(discharge, ...)
}
\arguments{
  \item{discharge}{Stream discharge during the experiment}
}
\value{
A new experiment object
}
\description{
Create a new experiment object based on the arguments defining the experiment
}
\examples{
}